refactor: :recycle: move non-critical utils to utils.py
